requirements.txt
flask
pymongo
nltk
spacy
transformers
language-tool-python
gunicorn  
pyspellchecker
textblob